refactor: improve unused tag
